move screens into lib/screens
