+++ lib/screens/full_screen_details.dart
