+++ lib/screens/homepage.dart
@@ -5,7 +5,7 @@ import 'package:callstats/routes/components/search_bottom_sheet.dart';
 import 'package:flutter/material.dart';
 import 'package:ionicons/ionicons.dart';
 
-import 'components/detailed_call_stat.dart';
+import '../routes/components/detailed_call_stat.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
